:violin: Remove sub-orchestrations to reduce complexity
Sub-orchestrations were introduced to allow index names to be retrieved
without running the reindexing but this is no longer required.

In addition, running multiple suborchestrations was failing after the
first and was proving difficult to resolve.
