--- ReIndexOrchestrator/function.json
@@ -1,9 +1,0 @@
-{
-  "bindings": [
-    {
-      "name": "context",
-      "type": "orchestrationTrigger",
-      "direction": "in"
-    }
-  ]
-}
